extra filter to its own function and verify behavior
the filter rules are going to get more complicated once we start encoding the hints Ryan has given.

This sets it up by extracting a boolean function to pass to filter. It's not really a method though and touches no internal state. Eventually we'll want more than one filter, filters that rely on internal state, and filters that return not booleans but floats on a range